+++ src/mltgnt/interfaces/__init__.py
@@ -1,6 +1,5 @@
 from mltgnt.interfaces.slack import SlackClientProtocol
 from mltgnt.interfaces.persona import PersonaProtocol
-from mltgnt.interfaces.chat import ChatPipelineProtocol
 from mltgnt.interfaces.types import (
     PersonaFMBase,
     Message,
@@ -20,7 +19,6 @@ from mltgnt.interfaces.turn import (
 __all__ = [
     "SlackClientProtocol",
     "PersonaProtocol",
-    "ChatPipelineProtocol",
     "PersonaFMBase",
     "Message",
     "ChatInput",